+++ src/main/java/uk/co/techblue/alfresco/dto/error/ServiceResponseStatus.java
@@ -18,79 +18,14 @@ package uk.co.techblue.alfresco.dto.error;
 import org.codehaus.jackson.annotate.JsonProperty;
 
 import uk.co.techblue.alfresco.dto.BaseDto;
+import uk.co.techblue.alfresco.dto.BaseStatus;
 
 /**
  * The Class ServiceResponseStatus.
  */
-public class ServiceResponseStatus extends BaseDto {
+public class ServiceResponseStatus extends BaseStatus {
 
     /** The Constant serialVersionUID. */
     private static final long serialVersionUID = 3517541718333984366L;
-
-    /** The code. */
-    @JsonProperty
-    private int code;
-
-    /** The name. */
-    @JsonProperty
-    private String name;
-
-    /** The description. */
-    @JsonProperty
-    private String description;
-
-    /**
-     * Gets the code.
-     * 
-     * @return the code
-     */
-    public int getCode() {
-        return code;
-    }
-
-    /**
-     * Sets the code.
-     * 
-     * @param code the new code
-     */
-    public void setCode(final int code) {
-        this.code = code;
-    }
-
-    /**
-     * Gets the name.
-     * 
-     * @return the name
-     */
-    public String getName() {
-        return name;
-    }
-
-    /**
-     * Sets the name.
-     * 
-     * @param name the new name
-     */
-    public void setName(final String name) {
-        this.name = name;
-    }
-
-    /**
-     * Gets the description.
-     * 
-     * @return the description
-     */
-    public String getDescription() {
-        return description;
-    }
-
-    /**
-     * Sets the description.
-     * 
-     * @param description the new description
-     */
-    public void setDescription(final String description) {
-        this.description = description;
-    }
 
 }
